board/beaglebone: add a genimage config
This patch adds a "genimage.cfg" genimage config file for the BeagleBone
(classic and black). The final generated "sdcard.img" image is ready to
use and contains the setup explained in the board readme.txt file: a
bootable vfat partition with the kernel, bootloader and DTB files, and
an ext2 rootfs partition.

The defconfig and readme file have been slightly changed to enable this
feature.

Tested on a BeagleBone Black.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>