board/wandboard: add a genimage config
This patch adds a "genimage.cfg" genimage config file for the Wandboard.
The final generated "sdcard.img" image is ready to use and contains the
setup explained in the board readme.txt file: U-Boot, its environment,
and an Ext4 rootfs partition.

The defconfig and readme file have been slightly changed to enable this
feature.

Tested on a Wandboard Solo.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>